Add files via upload
The task is to provide detail insights into the different marketing campaigns in a market.
a). Overview of the entire market’s development and the different campaigns and summarize the most important findings.

b). Assess the development of the quality of traffic. How is the overall development and how does each campaign evolve?

c). Advice a business developer who wants to spend an additional 250€ per week from week 31 onwards in a market campaign.

• Advice a business developer in which campaign to invest and why?

• How does this impact the overall performance in the market from week 31 onwards?